pyproject.toml
[tool.poetry]
name = "langchain-02"
version = "0.1.0"
description = ""
authors = ["pauloheck <pauloheck@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.1.0"
openai = "^1.6.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
